Create CCR_Wikdata_no_mappings_v1.1.csv
This csv contains the 529 used concepts for which a mapping was not found. The column "toDo_20251016" indicates if there is no mapping at all, or if there was a mapping but it was considered not valid. The discarded mapping is in the column "possible_mapping_to_check_again". I added this in case it's useful to reconsider these, or to help in investigating what the issues are.